FV; Wrap the TrieArena with a SimplexTrie.
This change introduces a SimplexTrie backed by a TrieArena. Of note, the
code that tests this version of the SimplexTrie duplicates the testing
done via the ptr version. At some point a trait should be added and the
testing of the two should be done with a single block of code.